# [chore] Improve logs bytes batch