@bogdandrutu 2025-03-03
@codeboten where is your comment coming from https://github.com/open-telemetry/opentelemetry-collector/pull/12529#pullrequestreview-2654548735?