@codeboten 2025-03-03
@bogdandrutu i just used the numbers in the description 😄