@codeboten 2025-03-03
Inverted them, the numbers make more sense now, thanks!

```
benchstat /tmp/before.txt /tmp/after1.txt
                                                       │ /tmp/before.txt │           /tmp/after1.txt            │
                                                       │     sec/op      │    sec/op     vs base                │
SplittingBasedOnItemCountManyLogsSlightlyAboveLimit-12      23.39m ± ∞ ¹   23.08m ± ∞ ¹       ~ (p=1.000 n=1) ²
SplittingBasedOnByteSizeManyLogsSlightlyAboveLimit-12       38.35m ± ∞ ¹   32.85m ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                                     29.95m         27.53m        -8.06%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                                       │ /tmp/before.txt │            /tmp/after1.txt            │
                                                       │      B/op       │     B/op       vs base                │
SplittingBasedOnItemCountManyLogsSlightlyAboveLimit-12     41.51Mi ± ∞ ¹   41.51Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
SplittingBasedOnByteSizeManyLogsSlightlyAboveLimit-12      41.60Mi ± ∞ ¹   41.58Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                                    41.56Mi         41.55Mi        -0.02%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                                       │ /tmp/before.txt │           /tmp/after1.txt            │
                                                       │    allocs/op    │  allocs/op    vs base                │
SplittingBasedOnItemCountManyLogsSlightlyAboveLimit-12      700.5k ± ∞ ¹   700.5k ± ∞ ¹       ~ (p=1.000 n=1) ²
SplittingBasedOnByteSizeManyLogsSlightlyAboveLimit-12       702.4k ± ∞ ¹   702.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                                     701.4k         701.3k        -0.02%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```